backend/api/utils/mixins.py
from rest_framework import mixins, viewsets


class RecipeMixins(
    mixins.ListModelMixin, mixins.CreateModelMixin,
    mixins.UpdateModelMixin, mixins.DestroyModelMixin,
    viewsets.GenericViewSet,
):
    pass
