backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.1
psycopg2-binary==2.9.5
django-filter==23.2
Pillow==10.0.0
python-decouple==3.5.0
gunicorn==20.1.0